🐞 BugFix - API 연동 (#80)
* IGS-47 BugFix: Time null Error 해결

* IGS-47 BugFix: API ERROR

* IGS-47 fix: UserOwnerJobPostingId를 DocumentId로 사용하고있던 버그 수정

* IGS-47 BugFix: API ERROR

* IGS-47 BugFix: API ERROR

* IGS-47 BugFix: API ERROR

* IGS-47 fix: 유저 검증 로직 및 유저가 지원한 공고에 대한 검증로직 추가 / 불필요 import 제거 / 주석 추가

* IGS-47 BugFix: API ERROR

* IGS-47 fix: 시간 관련 String 반환을 DateTimeUtil의 convert 메소드의 사용으로 전부 변경

* IGS-47 BugFix: API ERROR

* IGS-47 fix: UserOwnerJobPosting에 Notification에 관한 OneToMany 맵핑 추가

* IGS-47 BugFix: API ERROR

* IGS-47 fix: usecase 주석 추가

* IGS-47 BugFix: API ERROR

* IGS-47 BugFix: API ERROR

* IGS-47 BugFix: API ERROR

---------

+++ src/main/java/com/inglo/giggle/posting/application/service/UpdateOwnerJobPostingService.java
@@ -11,7 +11,6 @@ import com.inglo.giggle.core.utility.DateTimeUtil;
 import com.inglo.giggle.core.utility.S3Util;
 import com.inglo.giggle.posting.application.dto.request.UpdateOwnerJobPostingRequestDto;
 import com.inglo.giggle.posting.application.usecase.UpdateOwnerJobPostingUseCase;
-import com.inglo.giggle.posting.domain.CompanyImage;
 import com.inglo.giggle.posting.domain.JobPosting;
 import com.inglo.giggle.posting.domain.service.CompanyImageService;
 import com.inglo.giggle.posting.domain.service.JobPostingService;
@@ -19,6 +18,9 @@ import com.inglo.giggle.posting.domain.service.PostWorkDayTimeService;
 import com.inglo.giggle.posting.repository.mysql.CompanyImageRepository;
 import com.inglo.giggle.posting.repository.mysql.JobPostingRepository;
 import com.inglo.giggle.posting.repository.mysql.PostingWorkDayTimeRepository;
+import com.inglo.giggle.security.domain.mysql.Account;
+import com.inglo.giggle.security.domain.service.AccountService;
+import com.inglo.giggle.security.repository.mysql.AccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -32,6 +34,9 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class UpdateOwnerJobPostingService implements UpdateOwnerJobPostingUseCase {
 
+
+    private final AccountRepository accountRepository;
+    private final AccountService accountService;
 
     private final OwnerRepository ownerRepository;
     private final JobPostingRepository jobPostingRepository;
@@ -49,13 +54,22 @@ public class UpdateOwnerJobPostingService implements UpdateOwnerJobPostingUseCas
     @Transactional
     public void execute(List<MultipartFile> image, UUID accountId, Long jobPostingId, UpdateOwnerJobPostingRequestDto requestDto) {
 
+        // Account 조회
+        Account account = accountRepository.findById(accountId)
+                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_RESOURCE));
+
+        // 계정 타입 유효성 검사
+        accountService.checkOwnerValidation(account);
+
         // 고용주 조회
-        Owner owner = ownerRepository.findById(accountId)
-                .orElseThrow(() -> new CommonException(ErrorCode.INVALID_ACCOUNT_TYPE));
+        Owner owner = (Owner) account;
 
         // 공고 조회
         JobPosting jobPosting = jobPostingRepository.findById(jobPostingId)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_RESOURCE));
+
+        // 공고 수정 유효성 검사
+        jobPostingService.validateUpdateJobPosting(jobPosting, owner);
 
         // 주소 생성
         Address address = addressService.createAddress(
@@ -79,7 +93,7 @@ public class UpdateOwnerJobPostingService implements UpdateOwnerJobPostingUseCas
                 requestDto.workPeriod(),
                 requestDto.recruitmentNumber() == null ? null : requestDto.recruitmentNumber(), // null인 경우 모집인원 무관
                 requestDto.gender(),
-                requestDto.ageRestriction(),
+                requestDto.ageRestriction() == null ? null : requestDto.ageRestriction(),
                 requestDto.educationLevel(),
                 requestDto.visa(),
                 requestDto.recruiterName(),
@@ -94,6 +108,8 @@ public class UpdateOwnerJobPostingService implements UpdateOwnerJobPostingUseCas
         if (requestDto.workDayTimes() != null) {
             postingWorkDayTimeRepository.deleteAll(jobPosting.getWorkDayTimes());
 
+            jobPosting.getWorkDayTimes().clear();
+
             requestDto.workDayTimes().forEach(workDayTime -> postWorkDayTimeService.createPostingWorkDayTime(
                     workDayTime.dayOfWeek(),
                     (workDayTime.workStartTime() == null || workDayTime.workStartTime().isBlank()) ? null : DateTimeUtil.convertStringToLocalTime(workDayTime.workStartTime()),
@@ -102,7 +118,7 @@ public class UpdateOwnerJobPostingService implements UpdateOwnerJobPostingUseCas
             ));
         }
 
-        if (!requestDto.deletedImgIds().isEmpty()) {
+        if (requestDto.deletedImgIds() != null && !requestDto.deletedImgIds().isEmpty()) {
             companyImageRepository.findAllById(requestDto.deletedImgIds())
                     .forEach(companyImage -> s3Util.deleteFile(
                             companyImage.getImgUrl(),
